+++ docs/news.rst
@@ -7,6 +7,8 @@ Development (master)
 Features:
 
   - CheetahDirOwner caches compiled template in the template directory.
+  - CheetahDirOwner and DirOwner byte-compile compiled templates
+    to .pyc/.pyo. Errors on writing are silently ignored.
 
 Minor features: